Mann–Kendall Trend Test

Performs the Mann–Kendall trend test on a numeric vector to detect the presence of a monotonic trend (increasing or decreasing) over time. The test is nonparametric and accounts for tied observations in the data. The null hypothesis assumes there is no monotonic trend.

Usage

eda_mk_test(data, alpha = 0.05, quiet = TRUE)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

quiet

Logical scalar. If FALSE, prints a summary of of the statistical test to the console. Default is TRUE.

Value

A list containing the test results, including:

data: The data argument.

s_statistic: The Mann–Kendall test statistic \(S\).

s_variance: The variance of the test statistic under the null hypothesis.

p_value: The p-value associated with the two-sided hypothesis test.

reject: Logical. If TRUE, the null hypothesis is rejected at alpha.

msg: A character string summarizing the result, printed if quiet = FALSE.

Details

The statistic \(S\) is computed as the sum over all pairs \(i < j\) of the sign of the difference \(x_j - x_i\). Ties are explicitly accounted for when calculating the variance of \(S\), using grouped frequencies of tied observations. The test statistic \(Z\) is then computed based on the sign and magnitude of \(S\), and the p-value is derived from the standard normal distribution.

References

Kendall, M. (1975). Rank Correlation Methods. Griffin, London, 202 pp.

Mann, H. B. (1945). Nonparametric Tests Against Trend. Econometrica, 13(3): 245-25

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
eda_mk_test(data)
#> $data
#>   [1] 116.11412  90.69241 103.37862 105.22782  82.64148 109.39183 105.45024
#>   [8]  84.75166  94.25823 111.23967  86.69933  97.37938  96.58283 111.08042
#>  [15] 114.25324  99.55259  98.79250 106.57130 110.35033  88.61991  93.08293
#>  [22] 100.83571 104.75665 102.37172  92.25124 103.24165 108.40478 116.61977
#>  [29] 111.01600 108.00191 104.89789 108.15966  75.26442  89.19149 105.21227
#>  [36]  86.71029 118.16608 102.27068  87.32682 106.03875  88.81822 103.22338
#>  [43] 102.98355 114.61274 119.20752  89.24305  93.12479 106.15612  99.96043
#>  [50] 100.60333  97.72133  91.09217  99.29099  87.82165  96.03174 116.78254
#>  [57] 119.72330  98.41525  96.38403  92.93475 108.29597  92.21121  96.49941
#>  [64]  94.03990  95.46888  87.32682  93.65167 106.19333 105.58708  99.05894
#>  [71]  99.19472 100.19550  97.85936 107.34841  99.84156  88.23479  97.79369
#>  [78] 115.64516 111.22041 104.63076 106.33357  95.12631  89.60322  75.68793
#>  [85] 102.59373  96.36258 110.89977 107.88251 105.74549  98.24206 100.42963
#>  [92] 101.62222  97.24310  99.82175 102.24486  83.05105  98.18668 108.51155
#>  [99] 105.74684 107.53031
#> 
#> $s_statistic
#> [1] -60
#> 
#> $s_variance
#> [1] 112750
#> 
#> $p_value
#> [1] 0.8605227
#> 
#> $reject
#> [1] FALSE
#> 
#> $msg
#> [1] "\n - The Mann-Kendall test had a p-value of 0.861.\n - At a significance level of 0.05, we fail to reject the null hypothesis.\n - Therefore, there is NO evidence of a monotonic trend."
#>