Mann–Kendall–Sneyers Test for Change Point Detection

Performs the Mann–Kendall–Sneyers (MKS) test to detect the beginning of a monotonic trend in annual maximum series data. The test computes normalized progressive and regressive Mann–Kendall statistics and identifies statistically significant crossing points, indicating potential change points in the trend.

Usage

eda_mks_test(data, years, alpha = 0.05, quiet = TRUE)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

quiet

Logical scalar. If FALSE, prints a summary of of the statistical test to the console. Default is TRUE.

Value

A list containing the test results, including:

data: The data argument.

years: The years argument.

s_progressive: Normalized progressive Mann–Kendall-Sneyers statistics.

s_regressive: Normalized regressive Mann–Kendall-Sneyers statistics.

bound: Critical confidence bound for significance based on alpha.

crossing_df: Crossing points, including indices, years, and test statistics.

change_df: Subset of crossing_df with statistically significant crossings.

p_value: Two-sided p-value derived from the maximum crossing statistic.

reject: Logical. If TRUE, the null hypothesis of no change point is rejected.

msg: Character string summarizing the test result (printed if quiet = FALSE).

Details

The function computes progressive and regressive Mann–Kendall statistics \(S_t\), normalized by their expected values and variances under the null hypothesis. The crossing points where the difference between these normalized statistics changes sign are identified using linear interpolation. The significance of detected crossings is assessed using quantiles of the normal distribution.

References

Sneyers, R. (1990). On the statistical analysis of series of observations. Technical note No. 143, World Meteorological Organization, Geneva.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
eda_mks_test(data, years)
#> $data
#>   [1]  85.09204  98.48259  96.75074 111.55520  98.43460 110.63326 102.35489
#>   [8] 120.92595  83.07251  89.94641 116.74478  95.62255 113.27158 114.27544
#>  [15]  84.60421 111.82677 106.54959 113.19297  95.98798  94.25990  81.96696
#>  [22] 100.44151 100.24345 111.49112  97.26690  94.35158 107.04075 105.20372
#>  [29] 107.43372 101.73053 109.10044  93.58330  88.93986  92.14111 105.30740
#>  [36] 112.18120  84.70178 105.51114  94.83542  96.01779 103.24808 107.26851
#>  [43]  96.57706  94.41080  96.02413  99.10775 110.40952 102.40919  96.55350
#>  [50]  91.02249  98.98338  92.13937  90.92910 114.08476 102.50388  94.02291
#>  [57]  94.74065  94.91006  92.55806  91.13979 118.16085  83.30854 111.22617
#>  [64] 107.17357  96.59280  99.14638 103.38835  96.73917 101.31666  98.82788
#>  [71] 108.98828 110.77115 107.69158  85.54915  85.71614  97.22417 101.68775
#>  [78] 109.77722 102.96900  97.91774  96.94300 102.01555  89.55891 110.86274
#>  [85]  85.63202 104.14338  97.27396 109.76791  93.27818 107.34289 103.56664
#>  [92] 101.41236 103.49810 109.90516  95.04841 101.62542 115.25917 102.32179
#>  [99]  94.17772  98.77404
#> 
#> $years
#>   [1] 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915
#>  [16] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#>  [31] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#>  [46] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#>  [61] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#>  [76] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>  [91] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000
#> 
#> $s_progressive
#>   [1]  0.000000000  1.000000000  0.522232968  1.358732441  0.979795897
#>   [6]  1.315071011  1.351690671  1.979486637  0.834057656  0.268328157
#>  [11]  0.856348839  0.411435289  0.854124427  1.259132473  0.544358825
#>  [16]  0.810405304  0.823852555  1.098453328  0.734697000  0.324442842
#>  [21] -0.301969221 -0.253781368 -0.237694107  0.000000000 -0.140129810
#>  [26] -0.418789464 -0.229316010 -0.118539117  0.075032247  0.089205155
#>  [31]  0.288939088 -0.064865813 -0.433843555 -0.726397723 -0.582258714
#>  [36] -0.217934306 -0.601628535 -0.465160477 -0.641134244 -0.745666212
#>  [41] -0.651452411 -0.444332963 -0.554667287 -0.748455410 -0.841279499
#>  [46] -0.823729207 -0.559397660 -0.479954179 -0.586148819 -0.861580350
#>  [51] -0.852832167 -1.104750685 -1.365385099 -0.977312631 -0.863885270
#>  [56] -1.060130239 -1.197789713 -1.308057256 -1.510621544 -1.734796427
#>  [61] -1.331703750 -1.646076163 -1.370081699 -1.170311681 -1.188890886
#>  [66] -1.134491337 -1.001131697 -1.026919897 -0.953031061 -0.937886780
#>  [71] -0.719726334 -0.476402214 -0.285740204 -0.564675074 -0.827946905
#>  [76] -0.852123612 -0.782898632 -0.565199731 -0.469899066 -0.498559490
#>  [81] -0.546506476 -0.476530197 -0.719686431 -0.479022813 -0.736325519
#>  [86] -0.615424633 -0.641562445 -0.439699360 -0.616628533 -0.449597451
#>  [91] -0.346254058 -0.310297708 -0.212404716 -0.009798813 -0.138248896
#>  [96] -0.107617931  0.174532479  0.217920290  0.045347031  0.017868701
#> 
#> $s_regressive
#>   [1] -0.01786870  0.25092024  0.29158349  0.37399817  0.13927026  0.18326016
#>   [7] -0.02939644 -0.08628942 -0.39799054 -0.09941948  0.14986582 -0.15592905
#>  [13] -0.02883274 -0.31894819 -0.63034402 -0.33400333 -0.62582013 -0.79834069
#>  [19] -1.11724307 -1.01144482 -0.84755113 -0.52916562 -0.54794172 -0.56298329
#>  [25] -0.87903278 -0.83709549 -0.66734327 -0.85722061 -1.03058438 -1.26572562
#>  [31] -1.35359876 -1.62636822 -1.45039202 -1.17430042 -0.94633180 -1.14359980
#>  [37] -1.50634177 -1.18028683 -1.40311289 -1.24458294 -1.10975948 -1.27520000
#>  [43] -1.52271281 -1.43184058 -1.25802122 -1.12522871 -1.14144147 -1.48811634
#>  [49] -1.64134387 -1.53509790 -1.26309352 -1.27573566 -1.01323660 -0.70612492
#>  [55] -1.12671007 -1.27170157 -1.03165475 -0.80583738 -0.57438163 -0.26956651
#>  [61]  0.06990621 -0.39919679  0.06285952 -0.39236644 -0.65380292 -0.38343867
#>  [67] -0.31131331 -0.46483238 -0.19459744 -0.11897492 -0.01784103 -0.33764511
#>  [73] -0.82977382 -1.22996769 -0.72737117 -0.23354968  0.04960881  0.07923137
#>  [79] -0.36657309 -0.54354460 -0.38933141 -0.10495671 -0.11363310  0.49431153
#>  [85] -0.09004503  0.64333316  0.38321423  0.85412443  0.41143529  1.32344821
#>  [91]  0.98386991  0.62554324  0.98974332  0.75093926  0.18786729  0.97979590
#>  [97]  1.35873244  0.52223297 -1.00000000  0.00000000
#> 
#> $bound
#> [1] 1.959964
#> 
#> $crossing_df
#>   cross year  statistic       max
#> 1    53 1953 -1.1899384  90.92910
#> 2    55 1955 -0.9197108 102.50388
#> 3    57 1957 -1.1372399  94.74065
#> 4    73 1973 -0.4376558 107.69158
#> 5    75 1975 -0.7933734  85.71614
#> 6    80 1980 -0.4898663  97.91774
#> 7    81 1981 -0.5092288  96.94300
#> 8    99 1999  0.1790096  94.17772
#> 
#> $change_df
#> [1] cross     year      statistic max      
#> <0 rows> (or 0-length row.names)
#> 
#> $p_value
#> [1] 0.2340706
#> 
#> $reject
#> [1] FALSE
#> 
#> $msg
#> [1] "\n - The Mann-Kendall-Sneyers test had a p-value of 0.234.\n - At a significance level of 0.05, we fail to reject the null hypothesis.\n - Therefore, there is NO evidence of change point(s)."
#>