Wald–Wolfowitz Runs Test for Randomness

Applies the Wald–Wolfowitz runs test to a numeric vector in order to assess whether it behaves as a random sequence. The test statistic and p-value is computed using the number of runs (sequences of values above or below the median). Under the null hypothesis, the data is random. The runs test can be used to assess whether the residuals of a nonstationary model are random, indicating the suitability of the assumed nonstationary structure (e.g., linear).

Usage

eda_runs_test(values, years, alpha = 0.05)

Arguments

values

A numeric vector of values to check for randomness.

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

Value

A list containing the test results, including:

values: The values argument.

years: The years argument.

alpha: The significance level as specified in the alpha argument.

null_hypothesis: A string describing the null hypothesis.

alternative_hypothesis: A string describing the alternative hypothesis.

n: The length of the input vector after removing the median.

runs: The number of runs in the transformed sequence of residuals.

statistic: The runs test statistic, computed using runs and n.

p_value: The p-value derived from the normally distributed test statistic.

reject: If TRUE, the null hypothesis was rejected at significance alpha.

References

Wald, A. and Wolfowitz, J. (1940). On a test whether two samples are from the same population. Annals of Mathematical Statistics, 11(2), 147–162.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
sens_trend <- eda_sens_trend(data, years)
eda_runs_test(sens_trend$residuals, years)
#> $values
#>   [1]   3.4591070  -0.2548210   4.0780276  -5.7121758  -4.1019963   5.8170220
#>   [7]  18.3759451   7.8448449 -19.3704826   3.1951557  -4.8548835 -14.0041231
#>  [13]  15.8960238   9.7043435   7.7599675  -2.6344192  12.4022537   5.2769517
#>  [19]   5.9751249  -0.3766194  -8.9044044   0.4777749  -2.8826902  -7.0580551
#>  [25]  -6.4190872   4.2534043   0.2057172  14.2050622 -13.0367103  -2.4429078
#>  [31]  -6.3697003 -13.2841977  -5.0320028 -11.9802453  -3.8846573  -5.7262198
#>  [37]   4.5316228   7.6894095  10.4028158  10.9720948  -3.2556775   2.3393429
#>  [43] -10.4254574   8.8172799   1.9941589  -4.5579976  20.0140418   3.7121135
#>  [49]  -2.5697982  -9.5932598  -9.0913520 -16.7248099 -12.6076036   1.1704120
#>  [55]  14.2431448   5.5670667  -2.7751031  -7.6384031   0.4798927  18.3112193
#>  [61] -13.8913171  -5.2786273   9.3185863   7.5079721  -9.9864054   8.2432909
#>  [67]  -0.2057172   3.8273357  14.6473824  -0.8514123   3.0004590   7.0521799
#>  [73]  12.6558942  23.8521577 -11.5999526 -10.6746966 -11.1816573  -3.1532257
#>  [79]  -2.8515510   6.4116198 -14.4396657   3.4544708  -0.4803302  -6.9933186
#>  [85]   5.9791544   6.2065308  -7.1883797  16.3031722  -8.8611832   4.9651717
#>  [91]  13.5735386  23.8327449  18.3549657 -15.3155336  -4.2107676  20.1174146
#>  [97] -11.5633922   2.9811871  -7.6860026  -6.6695637
#> 
#> $years
#>   [1] 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915
#>  [16] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#>  [31] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#>  [46] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#>  [61] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#>  [76] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>  [91] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000
#> 
#> $alpha
#> [1] 0.05
#> 
#> $null_hypothesis
#> [1] "The input vector is random."
#> 
#> $alternative_hypothesis
#> [1] "The input vector is not random."
#> 
#> $n
#> [1] 100
#> 
#> $runs
#> [1] 50
#> 
#> $statistic
#> [1] -0.2010178
#> 
#> $p_value
#> [1] 0.8406846
#> 
#> $reject
#> [1] FALSE
#>