Spearman Test for Autocorrelation

Performs the Spearman serial correlation test on annual maximum series data to check for serial correlation at various lags. Reports the smallest lag where the serial correlation is not statistically significant at the given significance level (the least insignificant lag).

Usage

eda_spearman_test(data, alpha = 0.05)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

Value

A list containing the test results, including:

data: The data argument.

alpha: The significance level as specified in the alpha argument.

null_hypothesis: A string describing the null hypothesis.

alternative_hypothesis: A string describing the alternative hypothesis.

rho: Numeric vector of serial correlation estimates for lags \(1\) to \(n-3\).

least_lag: The smallest lag at which the serial correlation is insignificant.

significant: Indicates whether the serial correlation is significant at each lag.

reject: If TRUE, the null hypothesis was rejected at significance alpha.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
eda_spearman_test(data)
#> $data
#>   [1] 115.53949  97.62048  91.48151 103.99375  78.75553 109.02592 109.87513
#>   [8]  91.57034  75.35796  83.06758 111.02988 110.79918  97.81132 115.99110
#>  [15]  97.49414 114.96522 104.48972 101.82078  93.83436  86.25984  92.93219
#>  [22] 108.05278 101.79365 100.54119  84.71055 104.00399  96.18705 100.58256
#>  [29] 103.80280 101.53855 114.89921 104.92200 103.21021 101.94214 107.42460
#>  [36] 116.84394  92.50584 104.87693 100.38928  91.57844 110.66950 112.50907
#>  [43]  99.43985 111.01152 108.60272  94.35932  79.94742 102.48602 102.09311
#>  [50] 105.64982 104.92798  97.24127 102.44770 101.24951 102.63553  93.00900
#>  [57]  79.43679 102.37812  99.58353 106.73590  90.05232 106.93567 105.46208
#>  [64] 113.96973  95.44503 104.04250 103.99922  76.23319 104.59000 111.99336
#>  [71] 106.43052 102.14418 113.51713  86.05348  86.47162 106.82030  93.79462
#>  [78] 101.08264 111.23510  95.33119 103.83775  77.06348 117.92758  78.22464
#>  [85]  91.43614 109.47363 105.26578  81.73983 112.47440  99.33391 109.85445
#>  [92] 112.67286 100.08714  87.93782  95.52039  88.28031 107.00460 115.12215
#>  [99] 101.99070  93.67655
#> 
#> $alpha
#> [1] 0.05
#> 
#> $null_hypothesis
#> [1] "The least insignificant lag is 1."
#> 
#> $alternative_hypothesis
#> [1] "The least insignificant lag is greater than 1."
#> 
#> $rho
#>  [1] -0.088620903 -0.124393525 -0.013412582 -0.135349973 -0.056187010
#>  [6]  0.128158075 -0.026707648 -0.108564647 -0.045548654  0.208494876
#> [11] -0.144603337 -0.053661372 -0.036104833 -0.091806217 -0.047058824
#> [16]  0.060443454 -0.076445153 -0.039170231  0.236833785  0.088185654
#> [21] -0.068597858  0.107424221 -0.178978916 -0.204812030  0.098890469
#> [26] -0.151958534  0.031685795  0.336356036  0.124346076 -0.154614644
#> [31]  0.077968579 -0.023743177 -0.218173837  0.074793863  0.114597902
#> [36] -0.162042125  0.122503840  0.277731611 -0.141882602 -0.176382328
#> [41]  0.039625950 -0.038973823 -0.132745657  0.055502392 -0.113492063
#> [46] -0.290032399  0.280519271  0.296337403 -0.068506787  0.026458583
#> [51] -0.096632653 -0.169778550  0.094935245 -0.024360160  0.049407115
#> [56]  0.115715292  0.180761099 -0.138157362  0.053832753  0.006754221
#> [61] -0.100000000 -0.049786629  0.079895685 -0.175546976  0.156022409
#> [66]  0.023682200  0.075200535 -0.348240469  0.267741935 -0.155951057
#> [71] -0.171921182  0.200875753 -0.131257631 -0.187008547  0.523076923
#> [76]  0.039130435  0.015810277 -0.169960474  0.183116883  0.037593985
#> [81] -0.108771930 -0.011351909 -0.284313725 -0.150000000  0.496428571
#> [86]  0.221978022 -0.109890110 -0.188811189 -0.463636364  0.127272727
#> [91]  0.850000000  0.142857143 -0.535714286 -0.257142857 -0.400000000
#> [96]  0.000000000  1.000000000
#> 
#> $least_lag
#> [1] 1
#> 
#> $significant
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [13] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [25] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [37] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#> [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [73] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [85] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [97]  TRUE
#> 
#> $reject
#> [1] FALSE
#>