L-Moments Parameter Estimation for the Kappa Distribution

This functions estimates the parameters of the four-parameter Kappa distribution using the method of L-moments. Since there is no known closed form solution for the parameters in terms of the L-moments, the parameters are computed numerically using Newton-Raphson iteration.

Usage

fit_lmom_kappa(data)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

Value

A list containing the results of parameter estimation:

method: "L-moments".

params: numeric vector of 4 parameters in the order location, scale, shape (2).

Details

First, the sample L-moments of the data are computed using the lmom_sample() method. Then, the stats::optim() function is used to determine the parameters by minimizing the euclidian distance between the sample and theoretical L-moment ratios. The implementation of this routine is based on the deprecated homtest package, formerly avilable at https://CRAN.R-project.org/package=homtest.

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmom_kappa(data)
#> $method
#> [1] "L-moments"
#> 
#> $params
#> [1] 97.64945854 10.02823515  0.32276353  0.03890407
#>