L-Moments Parameter Estimation

For S-FFA only: Estimates the parameters of a stationary probability model using the L-moments.

Usage

fit_lmoments(data, distribution)

Arguments

data

Numeric vector of observed annual maximum series values. Must be strictly positive, finite, and not missing.

distribution

A three-character code indicating the distribution family. Must be "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

Value

A list containing the results of parameter estimation:

data: The data argument.

distribution: The distribution argument.

method: "L-moments".

params: Numeric vector of estimated parameters.

Details

First, the sample L-moments of the data are computed using utils_sample_lmoments(). Then, formulas from Hosking (1997) are used to match the parameters to the sample L-moments. The distributions "GNO", "PE3", and "LP3" use a rational approximation of the parameters since no closed-form expression is known.

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmoments(data, "GUM")
#> $data
#>   [1] 120.00224 103.56648  97.37379  91.94665 108.82779  82.40831  94.94096
#>   [8]  90.56819 102.17673  92.68039  98.05867  99.08948 100.12524 115.71024
#>  [15]  96.13517 111.17732  99.43153  90.23783  79.33490 104.12935 107.77373
#>  [22] 107.74647  90.52324 104.05805 117.28742  90.72178 100.51341  95.89050
#>  [29]  98.17972 102.58520  96.10162 100.28579 103.75042 102.36861 106.81857
#>  [36]  99.87023  93.63866 103.16125 101.24585 114.56949  98.74122 108.01745
#>  [43]  88.36622  92.08586  94.85833  91.17829  79.81302 106.05510  87.84966
#>  [50]  94.28917 115.41705 104.91905 115.68214  86.67678 102.79631  92.60434
#>  [57] 119.73140 117.24102 108.36749 103.51559  79.13414 101.57273 110.01306
#>  [64]  94.95882  94.36551  98.49159 114.57832  82.77162  88.09108  95.81234
#>  [71]  79.14392  84.59182  98.12339  97.29406  89.23987 112.77777 111.88500
#>  [78] 101.40517 109.86976  97.46669 114.10414  99.06067 119.42181 104.18372
#>  [85] 104.83997  95.58950  98.99218  96.55484 115.53452 106.02357  81.50147
#>  [92]  96.43393  86.41441  94.61247  99.85326 107.39677  97.63858 114.23139
#>  [99]  97.42237  87.23457
#> 
#> $distribution
#> [1] "GUM"
#> 
#> $method
#> [1] "L-moments"
#> 
#> $params
#> [1] 95.076997  8.214399
#>