Maximum Likelihood Parameter Estimation

Estimates parameters of a probability distribution with an optional nonstationary structure by maximizing the log‐likelihood. Initial values are obtained through L‐moment parameter estimation, and optimization is performed via stats::nlminb() with repeated perturbations if needed.

Usage

fit_maximum_likelihood(
  data,
  distribution,
  prior = NULL,
  years = NULL,
  structure = NULL
)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

distribution

Character scalar. A three-character code indicating a distribution family. Must be one of: "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

prior

Numeric vector of length 2. Specifies the Beta prior shape parameters \((p, q)\) for the shape parameter \(\kappa\). Only used when distribution = "GEV".

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

structure

Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A list containing the results of parameter estimation:

params: Numeric vector of estimated parameters.

mll: Maximum log‐likelihood value.

Details

Calls fit_lmom_fast() on data to obtain initial parameter estimates.

Initializes trend parameters to zero if necessary.

For WEI models, sets the location parameter to zero to ensure support.

Defines an objective function using loglik_fast() or general_loglik_fast().

Runs stats::nlminb() with box constraints. Attempts optimization up to 100 times if a maximum cannot be found.

Note

Although the more modern stats::optim() function is preferred over stats::nlminb(), we use stats::nlminb() because it supports infinite values of the likelihood function.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
structure <- list(location = TRUE, scale = FALSE)
fit_maximum_likelihood(data, "GNO", NULL, years, structure)
#> $method
#> [1] "MLE"
#> 
#> $params
#> [1] 99.45758050  1.59710972 11.41849812 -0.07884203
#> 
#> $mll
#> [1] -385.4173
#>