Evaluate Goodness-of-Fit for Fitted Flood Models

Computes multiple performance metrics and diagnostic indicators to assess the quality of a fitted flood frequency model. This includes accuracy (residual statistics), fitting efficiency (information criteria), and uncertainty (coverage-based metrics using confidence intervals).

Usage

model_diagnostics(
  data,
  distribution,
  params,
  uncertainty,
  years = NULL,
  structure = NULL,
  alpha = 0.05,
  pp_formula = "Weibull"
)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

distribution

Character scalar. A three-character code indicating a distribution family. Must be one of: "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

uncertainty

List; estimated reutrn levels and confidence intervals generated by uncertainty_bootstrap() or uncertainty_rfpl().

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

structure

Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

pp_formula

Character (1); string specifying the plotting position formula. Must be one of "Weibull", "Blom", "Cunnane", "Gringorten", or "Hazen"

Value

List containing model assessment metrics:

data: The data argument.

estimates: Quantile estimates for empirical return periods.

R2: Coefficient of determination from linear regression of estimates vs. data.

RMSE: Root mean squared error of quantile estimates.

bias: Mean bias of quantile estimates.

AIC: Akaike Information Criterion.

BIC: Bayesian Information Criterion.

AIC_MLL: Akaike Information Criterion, computed using MLE.

BIC_MLL: Bayesian Information Criterion, computed using MLE.

AW: Average width of the confidence interval(s).

POC: Percent of observations covered by the confidence interval(s).

CWI: Confidence width index, a metric that combines AW and POC.

See also

Examples

# Initialize example data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(100, 10)

# Perform uncertainty analysis
uncertainty <- uncertainty_bootstrap(data, "NOR", "L-moments")

# Evaluate model diagnostics
model_diagnostics(data, "NOR", params, uncertainty)
#> $data
#>   [1] 113.79449  86.25379  95.89541 101.91088 110.64448  92.80289  66.20745
#>   [8] 101.76545 105.14418  97.68805  83.51785 102.70968  86.82404 111.54073
#>  [15]  97.98395  99.96687  99.81895  95.15258  87.57367  95.40822  91.69506
#>  [22]  88.04409 101.20507 105.82843  93.86866 104.63807  97.09908 107.45821
#>  [29] 104.61719  93.12696  87.88099  88.90739  90.79209  92.38459  88.58345
#>  [36] 117.87419 107.94493  95.44703 102.17106  90.41650  69.52560 118.92147
#>  [43]  98.87746  95.92466 100.85646 102.19875  93.54169 102.07378 102.21039
#>  [50] 107.09135  96.45425 101.78812 125.52394  92.14230 101.96464 117.70845
#>  [57]  95.69500  97.02246  80.92915  99.49001  82.79532 130.00649 100.68805
#>  [64]  88.34959  96.66235 106.53280 114.47210  98.23928  95.30328  97.86018
#>  [71]  90.53086 107.57747  90.35654 108.57547  85.90441 100.53240 100.39905
#>  [78] 115.23569  91.24060 110.07792  97.39674  78.80304  93.12105 116.86685
#>  [85]  76.67217 109.59920  93.22673  99.73017  91.59067  77.74570  95.81638
#>  [92] 120.79266  94.38469  94.98544  87.66035  95.80415  71.69704  99.85433
#>  [99]  97.61136 104.81938
#> 
#> $estimates
#>   [1] 123.30079 120.57856 118.85177 117.55301 116.49673 115.59780 114.80973
#>   [8] 114.10420 113.46263 112.87214 112.32341 111.80947 111.32497 110.86568
#>  [15] 110.42824 110.00990 109.60838 109.22178 108.84850 108.48716 108.13657
#>  [22] 107.79571 107.46367 107.13967 106.82300 106.51302 106.20918 105.91097
#>  [29] 105.61792 105.32963 105.04569 104.76577 104.48953 104.21668 103.94693
#>  [36] 103.68003 103.41572 103.15377 102.89397 102.63612 102.38000 102.12543
#>  [43] 101.87224 101.62024 101.36926 101.11915 100.86973 100.62085 100.37236
#>  [50] 100.12409  99.87591  99.62764  99.37915  99.13027  98.88085  98.63074
#>  [57]  98.37976  98.12776  97.87457  97.62000  97.36388  97.10603  96.84623
#>  [64]  96.58428  96.31997  96.05307  95.78332  95.51047  95.23423  94.95431
#>  [71]  94.67037  94.38208  94.08903  93.79082  93.48698  93.17700  92.86033
#>  [78]  92.53633  92.20429  91.86343  91.51284  91.15150  90.77822  90.39162
#>  [85]  89.99010  89.57176  89.13432  88.67503  88.19053  87.67659  87.12786
#>  [92]  86.53737  85.89580  85.19027  84.40220  83.50327  82.44699  81.14823
#>  [99]  79.42144  76.69921
#> 
#> $R2
#> [1] 0.9772745
#> 
#> $RMSE
#> [1] 3.047736
#> 
#> $bias
#> [1] 2.079555
#> 
#> $AIC
#> [1] 115.4399
#> 
#> $BIC
#> [1] 120.6502
#> 
#> $AW
#> [1] 5.087545
#> 
#> $POC
#> [1] 97.95918
#> 
#> $CWI
#> [1] 4.795182
#> 
#> $MLL_AIC
#> [1] 770.7833
#> 
#> $MLL_BIC
#> [1] 775.9936
#>