Helper Function for Quantile Functions

A helper function used by quantile_xxx(). This function does not validate parameters and is designed for use in other methods.

Usage

quantile_fast(p, distribution, params, slice, structure)

Arguments

p

Numeric vector of probabilities between 0 and 1 with no missing values.

distribution

Character scalar. A three-character code indicating a distribution family. Must be one of: "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

slice

Numeric scalar specifying the year at which to evaluate the quantiles of a nonstationary probability distribution. The slice does not have to be an element of the years argument. Note that if structure$location and structure$scale are both FALSE, this argument will have no effect the output of the function.

structure

Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A numeric vector of quantiles with the same length as p.

See also

Examples

# Initialize p, params, and structure
p <- runif(n = 10)
params <- c(0, 1, 1, 0)
structure <- list(location = FALSE, scale = TRUE)

# Compute the log-likelihood in the year 2000
quantile_fast(p, "GEV", params, 2000, structure)
#>  [1]  1.6119281 -0.7093861 -1.2223080 -0.7742434  0.9181472 -0.8100189
#>  [7]  4.3439162  1.9114303  3.4545270  1.2281361