Quantile Functions for Probability Models

Compute the quantiles for stationary and nonstationary variants of nine different distributions: "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI". In total, these methods compute the quantiles for 36 different probability models.

Usage

quantile_gum(p, params, slice = 1900, structure = NULL)

quantile_nor(p, params, slice = 1900, structure = NULL)

quantile_lno(p, params, slice = 1900, structure = NULL)

quantile_gev(p, params, slice = 1900, structure = NULL)

quantile_glo(p, params, slice = 1900, structure = NULL)

quantile_gno(p, params, slice = 1900, structure = NULL)

quantile_pe3(p, params, slice = 1900, structure = NULL)

quantile_lp3(p, params, slice = 1900, structure = NULL)

quantile_wei(p, params, slice = 1900, structure = NULL)

quantile_kap(p, params, slice = 1900, structure = NULL)

Arguments

p

Numeric vector of probabilities between 0 and 1 with no missing values.

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

slice

Numeric scalar specifying the year at which to evaluate the quantiles of a nonstationary probability distribution. The slice does not have to be an element of the years argument. Note that if structure$location and structure$scale are both FALSE, this argument will have no effect the output of the function.

structure

Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A numeric vector of quantiles with the same length as p.

See also

Examples

# Initialize p and params
p <- runif(n = 10)
params <- c(0, 1, 0)

# Compute the quantiles
quantile_wei(p, params)
#>  [1]   0 Inf Inf   0   0   0   0 Inf   0   0