L-Distance Method for Distribution Selection

Selects a distribution from a set of candidate distributions by minimizing the Euclidean distance between the theoretical L-moment ratios \((\tau_3, \tau_4)\) and the sample L-moment ratios \((t_3, t_4)\).

NS-FFA: To select a distribution for a nonstationary model, include the observation years (ns_years) and the nonstationary model structure (ns_structure). Then, this method will detrend the data internally using the data_decomposition() function prior to distribution selection.

Usage

select_ldistance(data, ns_years = NULL, ns_structure = NULL)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

ns_years

NS-FFA only. Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

ns_structure

NS-FFA only. Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A list with the results of distribution selection:

method: "L-distance".

data: The data argument (S-FFA) or the detrended dataset (NS-FFA).

metrics: A list of L-distance metrics for each candidate distribution.

recommendation: The name of the distribution with the smallest L-distance.

Details

For each candidate distribution, this method computes the Euclidean distance between sample L-moment ratios (\(\tau_3\), \(\tau_4\)) and the closest point on the theoretical distribution's L-moment curve. For two-parameter distributions (Gumbel, Normal, Log-Normal), the theoretical L-moment ratios are compared directly with the sample L-moment ratios. The distribution with the minimum distance is selected. If a distribution is fit to log-transformed data (Log-Normal or Log-Pearson Type III), the L-moment ratios for the log-transformed sample are used instead.

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
select_ldistance(data)
#> $method
#> [1] "L-distance"
#> 
#> $data
#>   [1] 112.36210  90.96671  88.81809 105.49874 101.44455  91.36199 108.24594
#>   [8] 101.76501  92.67487 100.88402  94.50747 110.45014  87.42728 112.88892
#>  [15] 105.50069 113.28995 103.00665 118.53577 102.97076  98.57930 112.89908
#>  [22]  97.18895 111.13498 109.83232 109.55026  96.96021  97.91620  87.01459
#>  [29]  99.28443  93.81179  97.34613  93.89084 103.55019 109.96499  79.31277
#>  [36]  93.83550  91.45473 109.03163 115.78444 105.72620  92.92973 107.96002
#>  [43]  93.29243  94.34594  99.73184  90.12823  92.97532 109.12187  85.51126
#>  [50]  97.94028 100.59896  91.66640 106.53450  99.54317 109.12172  99.03369
#>  [57] 122.21154 109.30758  98.84861 109.11931  89.31727 107.16006  87.92649
#>  [64]  97.34781  94.12382 102.02802  95.21949 107.10338 108.55868 111.70885
#>  [71]  88.50561 113.18589  92.41219 115.19096  91.95496 115.27030  86.47308
#>  [78]  95.39673 117.46465  98.83503  91.59992 113.50751  90.05050  92.04249
#>  [85]  99.44339 110.29035 100.57330 103.17333  93.53476  82.46938  83.48580
#>  [92] 102.76144  93.08456  99.57173 109.96268 104.46125 102.70581 117.61106
#>  [99]  86.70633  92.06702
#> 
#> $metrics
#> $metrics$GUM
#> [1] 0.1743756
#> 
#> $metrics$NOR
#> [1] 0.08134432
#> 
#> $metrics$LNO
#> [1] 0.07437066
#> 
#> $metrics$GEV
#> [1] 0.0636822
#> 
#> $metrics$GLO
#> [1] 0.1204359
#> 
#> $metrics$GNO
#> [1] 0.07637753
#> 
#> $metrics$PE3
#> [1] 0.07601018
#> 
#> $metrics$LP3
#> [1] 0.07391115
#> 
#> $metrics$WEI
#> [1] 0.05763824
#> 
#> 
#> $recommendation
#> [1] "WEI"
#>