L-Kurtosis Method for Distribution Selection

Selects a probability distribution by minimizing the absolute distance between the theoretical L-kurtosis (\(\tau_4\)) and the sample L-kurtosis (\(t_4\)). Only supports 3-parameter distributions.

Usage

select_lkurtosis(data)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

Value

A list with the results of distribution selection:

method: "L-kurtosis".

data: The data argument.

metrics: A list of L-kurtosis metrics for each distribution.

recommendation: Name of the distribution with the smallest L-kurtosis metric.

Details

This method computes the distance between the sample and theoretical L-kurtosis values at a fixed L-skewness. For three parameter distributions, the shape parameter that best replicates the sample L-skewness is determined using stats::optim().

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
select_lkurtosis(data)
#> $method
#> [1] "L-kurtosis"
#> 
#> $data
#>   [1] 102.56928  97.68225  72.17210 106.98132 101.15245  72.87185  98.46018
#>   [8] 101.75773  97.67156  90.50121 113.06182  95.83942 103.72385 101.94554
#>  [15]  94.65623  76.35925 107.50753 100.23578 112.95685 104.52338 108.84731
#>  [22] 100.58835 102.51379 103.85555  87.34236 103.05540  96.86775  96.07440
#>  [29]  94.67134  93.00802  73.18456 102.90506 117.04368  95.54740 105.50263
#>  [36]  83.87351 113.75235 103.23683  90.53634  99.54861  93.73578 100.86708
#>  [43] 101.32189 106.35173  93.95137 108.27911  97.68586 108.75957 101.26202
#>  [50] 102.42864 117.77877 106.00204  93.59048 117.69072 103.19826  94.83803
#>  [57] 100.08120 108.81954 119.15510  85.23162 103.83925  99.22607 109.92254
#>  [64]  92.66711 102.46391 108.62512 102.93538 103.78512 101.40673 105.30224
#>  [71] 114.79813 102.88134 108.42450  98.82199 107.07486 109.63430 105.38478
#>  [78]  95.13129  96.01258 116.88485 102.94479 100.76935  93.30077  89.89977
#>  [85]  90.65092  93.12266  91.76308 104.91010  95.36140  95.45133  77.20147
#>  [92]  91.00006  96.47586 103.08095 102.79732  95.16230 109.32038 107.69312
#>  [99]  97.39382 110.75419
#> 
#> $metrics
#> $metrics$GEV
#> [1] 0.09368644
#> 
#> $metrics$GLO
#> [1] 0.02331306
#> 
#> $metrics$GNO
#> [1] 0.06782938
#> 
#> $metrics$PE3
#> [1] 0.07245138
#> 
#> $metrics$LP3
#> [1] 0.08458052
#> 
#> $metrics$WEI
#> [1] 0.07170038
#> 
#> 
#> $recommendation
#> [1] "GLO"
#>