(Generalized) Regula-Falsi Confidence Intervals for Flood Quantile Estimates

Calculates return level estimates and confidence intervals at specified return periods (defaults to 2, 5, 10, 20, 50, and 100 years) using the Regula-Falsi profile likelihood or Regula-Falsi generalized profile likelihood root‐finding methods.

Usage

uncertainty_rfpl(
  data,
  distribution,
  prior = NULL,
  years = NULL,
  structure = NULL,
  slices = 1900,
  alpha = 0.05,
  eps = 0.01,
  periods = c(2, 5, 10, 20, 50, 100)
)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

distribution

Character scalar. A three-character code indicating a distribution family. Must be one of: "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

prior

Numeric vector of length 2. Specifies the Beta prior shape parameters \((p, q)\) for the shape parameter \(\kappa\). Only used when distribution = "GEV".

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

structure

Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

slices

Numeric vector specifying the years at which to evaluate the return levels confidence intervals of a nonstationary probability distribution. The slices do not have to be elements of the years argument.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

eps

Numeric scalar. The log-likelihood tolerance for the Regula-Falsi convergence (default is 0.01).

periods

Numeric vector used to set the return periods for FFA. All entries must be greater than or equal to 1.

Value

A list containing the following three items:

method: "Bootstrap"

structure: The value of the structure argument.

slices: A list of lists containing the results for each slice.

Each element of slices is a list with the following five items:

estimates: Estimated quantiles for each return period.

ci_lower: Lower bound of the confidence interval for each return period.

ci_upper: Upper bound of the confidence interval for each return period.

periods: Vector of return periods; defaults to c(2, 5, 10, 20, 50, 100).

year: The year at which the estimates were computed (nonstationary models only).

Details

Fits the distribution using fit_maximum_likelihood() to obtain parameter estimates and log‐likelihood.

Defines an objective function \(f(y_p, p)\) using the reparameterized log-likelihood function.

Iteratively brackets the root by rescaling initial guesses by 0.05 until \(f(y_p, p)\) changes sign.

Uses the Regula Falsi method to solve \(f(y_p, p) = 0\) for each return-period probability.

Returns lower and upper confidence bounds at significance level alpha given the return level estimates.

Note

Although the more modern stats::optim() function is preferred over stats::nlminb(), we use stats::nlminb() because it supports infinite values of the likelihood function.

RFPL uncertainty quantification can be numerically unstable for some datasets. If this function encounters an issue, it will return an error and recommend using the parametric bootstrap method (uncertainty_bootstrap) instead.

References

Vidrio-Sahagún, C.T., He, J. Enhanced profile likelihood method for the nonstationary hydrological frequency analysis, Advances in Water Resources 161, 10451 (2022). doi:10.1016/j.advwatres.2022.104151

Vidrio-Sahagún, C.T., He, J. & Pietroniro, A. Multi-distribution regula-falsi profile likelihood method for nonstationary hydrological frequency analysis. Stoch Environ Res Risk Assess 38, 843–867 (2024). doi:10.1007/s00477-023-02603-0

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
uncertainty_rfpl(data, "GLO")
#> $method
#> [1] "RFPL"
#> 
#> $structure
#> $structure$location
#> [1] FALSE
#> 
#> $structure$scale
#> [1] FALSE
#> 
#> 
#> $slices
#> $slices[[1]]
#> $slices[[1]]$estimates
#> [1] 101.1298 108.5625 112.8805 116.8398 121.8331 125.5208
#> 
#> $slices[[1]]$ci_lower
#> [1]  99.25391 106.54348 110.45048 113.82004 117.79865 120.55549
#> 
#> $slices[[1]]$ci_upper
#> [1] 102.9703 110.9292 116.1821 121.4801 128.8411 134.7545
#> 
#> $slices[[1]]$periods
#> [1]   2   5  10  20  50 100
#> 
#> $slices[[1]]$year
#> NULL
#> 
#> 
#>