Quantile Functions for Probability Models

Compute the quantiles for stationary and nonstationary probability models.

NS-FFA: To compute the quantiles for a nonstationary probability model, specify a time slice (ns_slice) and the nonstationary model structure (ns_structure).

Usage

utils_quantiles(p, distribution, params, ns_slice = 0, ns_structure = NULL)

Arguments

p

Numeric vector of probabilities between 0 and 1 with no missing values.

distribution

A three-character code indicating the distribution family. Must be "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

ns_slice

NS-FFA only. Numeric scalar specifying the year at which to evaluate the quantiles of a nonstationary probability distribution. ns_slice does not have to be an element of the ns_years argument.

ns_structure

NS-FFA only. Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A numeric vector of quantiles with the same length as p.

Examples

p <- runif(n = 100)
params <- c(0, 1, 0)
utils_quantiles(p, "GEV", params)
#>   [1]  0.90287503  1.25133683  3.26406638  1.17388889  0.99072515 -1.41229746
#>   [7] -0.46683505  1.04387896  2.32610708  1.30750670  0.12333376  0.68928581
#>  [13]  0.35947517 -0.19900579  0.08026637  0.18538955  1.03307935 -0.42077023
#>  [19] -0.91324564 -1.16420401  0.42143644  3.07340999 -0.29235503  0.69195019
#>  [25]  0.20717061  1.84805853  0.26978712  1.73183484  1.58463632  0.41230524
#>  [31]  1.20629030  0.70868657 -1.75826006  1.17189610  1.07705253 -0.33641397
#>  [37]  0.29054284 -0.15184243 -0.43592617 -0.95893732 -0.41557793 -0.02149417
#>  [43] -1.20762902 -0.53329041  0.41237733  0.50745928  2.40011084  0.27848776
#>  [49] -1.21763381  0.27557550  2.01454446  0.66474823 -0.14393968 -0.49081212
#>  [55]  3.85608935  0.81204611 -0.08313228  0.77680317  1.42389453  0.35191056
#>  [61] -0.41790688 -0.19594937 -0.49172035 -0.83419590  0.32854102  0.24563560
#>  [67]  2.94342599 -0.77920030  0.25142949  0.16690915  3.68633734 -0.19820451
#>  [73] -0.06434615  0.10101678  1.77977770 -0.06475094  0.65049125 -0.36510008
#>  [79]  0.17576545 -0.46011969  0.32580414 -0.62608574  0.97463001 -0.03395770
#>  [85]  0.28250578  1.87656141  2.54000127  0.65386897  3.33902292  0.36043039
#>  [91] -0.50790824 -0.35828696 -0.28485790  0.49760815  0.14835837 -0.10383544
#>  [97] -0.33839157  2.50034970 -0.38749960 -0.06384954