Block-Bootstrap Mann-Kendall Test for Trend Detection

Performs a bootstrapped version of the Mann-Kendall trend test to account for serial correlation in annual maximum series data. The procedure uses Spearman’s serial correlation test to estimate the least insignificant lag, then applies a bootstrap procedure to obtain the empirical p-value and confidence bounds for the Mann-Kendall S-statistic.

Usage

eda_bbmk_test(data, alpha = 0.05, samples = 10000L, quiet = TRUE)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

samples

Integer scalar. The number of bootstrap samples. Default is 10000.

quiet

Logical scalar. If FALSE, prints a summary of of the statistical test to the console. Default is TRUE.

Value

List; the results of the test, including:

data: The data argument.

s_bootstrap: Vector of bootstrapped Mann-Kendall S-statistics.

s_statistic: The Mann-Kendall S-statistic computed on the original series.

p_value: Empirical two-sided p-value derived from the bootstrap distribution.

bounds: Confidence interval bounds for the null distribution of the statistic.

reject: Logical. If TRUE, the null hypothesis was rejected at alpha.

msg: Character string summarizing the test result (printed if quiet = FALSE).

Details

The block size for the bootstrap is selected as least_lag + 1, where least_lag is estimated using eda_spearman_test(). Bootstrap samples are generated by resampling blocks of the original data without replacement and computing the Mann-Kendall S-statistic. This procedure removes serial correlation from the data.

References

Bayazit, M., 2015. Nonstationarity of hydrological records and recent trends in trend analysis: a state-of-the-art review. Environ. Processes 2 (3), 527–542. doi:10.1007/s40710-015-0081-7

Khaliq, M.N., Ouarda, T.B.M.J., Gachon, P., Sushama, L., St-Hilaire, A., 2009. Identification of hydrological trends in the presence of serial and cross correlations: a review of selected methods and their application to annual flow regimes of Canadian rivers. J. Hydrol. 368 (1–4), 117–130. doi:10.1016/j.jhydrol.2009.01.035

Sonali, P., Nagesh Kumar, D., 2013. Review of trend detection methods and their application to detect temperature changes in India. J. Hydrol. 476, 212–227. doi:10.1016/j.jhydrol.2012.10.034

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
eda_bbmk_test(data, samples = 1000L)
#> $data
#>   [1] 100.65487  89.01491  93.66822  79.36346 126.48932  88.46602  96.59362
#>   [8] 107.86363  87.29487 105.42142 100.75106 105.58514 104.15406  85.47700
#>  [15] 109.41206  96.61064  99.24426 100.40204 101.24301  90.01567 112.33390
#>  [22] 103.40424  95.27298 107.08753  84.71041 102.37425  86.87186 107.47029
#>  [29]  84.37482 100.71053  93.60465  91.54804 106.75245 111.53376  83.13495
#>  [36]  90.97185 113.17634 111.00190 112.03768  85.68729 113.82911 100.03126
#>  [43]  99.22113 104.41428 101.28923  91.69786  94.96407  88.06359  92.48277
#>  [50] 114.55841  91.71396 102.89774  95.19947  93.95171 114.60110 101.49679
#>  [57]  85.66679  99.89697  97.87764  90.93660  78.97848 118.93360  90.31874
#>  [64]  98.97397 102.39960 100.60899  78.22424  98.82140 101.12295 100.07886
#>  [71] 118.77744 121.58757 107.09715 107.66983  96.91789 110.12002  90.80948
#>  [78] 105.63380 103.22483 103.66674 111.29835  90.58502 102.17838 114.15412
#>  [85]  96.16267  98.25914  97.78255  89.90471 104.80725 116.04407  84.84975
#>  [92]  85.83976 108.76777 106.24132 121.12277  96.43876  89.35536 110.77117
#>  [99] 111.81576 101.98392
#> 
#> $s_bootstrap
#>    [1] -410 -276 -322 -136  538 1006  116  200  -44   80  148  234  -86   32
#>   [15]  360  146 -152  -28 -142   86 -134  550 -104 -148  166 -118  494 -124
#>   [29] -186  206  294 -338  262  270  148 -280 -174  522  342  106    2   78
#>   [43]    6  232 -146 -208  256  -58 -242  512   42  264  522 -218   88 -542
#>   [57] -244 -240   34  -42 -226  258  406 -300   80   76 -248  798 -642  434
#>   [71]  414 -138   34  616  500 -552  380 -568  286  110   44  334  390 -570
#>   [85]  266 -372  -42  434  516 -304  450  128  554   18 -356  738 -188 -374
#>   [99]  -96 -254  152  336  298 -532 -880   26 -128 -504 -442 -276  -28   36
#>  [113]  -76  -48  120  146  -94 -256  406   86 -180 -268  396  312  162  372
#>  [127] -372  336  -82  138   92 -146  576  -80  296   22  -76 -646  658 -532
#>  [141]  194 -214 -238  136  134  658 -156 -276  -54 -266 -508 -226  162  236
#>  [155]  148  236  112  106  488 -146  316 -562  190  428  694 -388  -10 -436
#>  [169]  256  710  610  106 -106   26   44  626  224 -264  112   14 -304   16
#>  [183]    2  -76  458  286 -256 -170 -526  448 -508  240  298  282  -36  728
#>  [197]  286   38 -344 -436 -430  348 -534 -270 -524   -2 -166  216 -406 -342
#>  [211]   80 -196  476  110  360  536  570 -204 -550 -348  154  288  158 -182
#>  [225] -136   20  370  288   74  164  404  -40 -224  708  230   44   48  204
#>  [239]  494 -324 -330 -138  -56 -132  116  528   20  240    4  -90 -350  -26
#>  [253] -952  274 -202  -98 -952  138 -442  490 -612   34  408  284 -248  122
#>  [267] -144  330 -104  690 -494 -444  -36 -936  196  230   82  390   92  -60
#>  [281]  152   66  104 -118 -150  -84   14 -420  252 -356  -26 -338   48 -276
#>  [295] -276  392  144  368 -176   20 -256  486  260  396 -382  562   16 -568
#>  [309] -418  -70  324  -66  310 -530  574  164  402 -340   46 -116   52 -192
#>  [323] -526  -12  386  -10  130  -50 -204  358  230  106 -544   54  320  276
#>  [337]  652 -528  -50  436 -192  134  -74 -130  -22 -482  330  230  408  324
#>  [351]  430  242  610  584 -132 -390  112 -518 -292  454  252   22  276 -230
#>  [365]  -70  422 -198  412  358  204 -386 -366 -540  360  152  178  594  170
#>  [379]  274 -420  194  -88 -692 -378 -358 -366 -480  118  -74   -6 -168   54
#>  [393]  132  398   80 -230  -84 -710 -358 -242  310  164  414   90  244 -478
#>  [407]  -64 -150  176 -360   72  280  -48 -176 -102  -64   68 -138   64 -224
#>  [421]  154   92  410  510 -114 -696 -206  310  302  172  100   98   84 -186
#>  [435]   36  180 -560    0  488 -104 -122 -238  624  308  184   28   54 -670
#>  [449]  626   14  -50  318   44 -160   34 -174  178 -848  600  -46  244  414
#>  [463] -296 -206  592 -298   86   56   88  530   42 -358 -500  214 -222  -20
#>  [477] -544  312  426  234 -254  316 -440 -366  360  -76 -156 -250   46 -452
#>  [491]   24   80  108  196  382 -470 -400  372 -360   90  576 -278  186 -368
#>  [505] -486 -224   -6  206 -406 -434 -312 -506 -872    8  218  -58 -132  132
#>  [519]  586 -102 -212  444   58    8 -622  344  298 -236  214 -232 -206   94
#>  [533] -148  256  606 -244 -144 -204  260 -174  158   32  538   54 -584 -406
#>  [547]  232   46  -10 -498   84  196 -446 -474  152   44 -290 -304  116 -286
#>  [561] -262 -190  110  340 -616  -50 -838 -416   96  130    8  192  292  312
#>  [575] -380  192   24  216 -490 -120  114  220  362   40  -78 -252    8  180
#>  [589]  194 -308   32  334 -208   52  -62  -12  420  468 -170 -180  144 -516
#>  [603]   14 -156   -6  700  388 -320 -316  552  -24   26  676 -116  -32 -596
#>  [617] -214 -652  644 -196 -276 -358  372 -428   72  -10   82  130    6  -96
#>  [631] -330  -68 -612  614 -310  712  -16 -114 -286 -316 -256 -138   76  506
#>  [645]  412 -328   10    6  100  600 -164  -40  546 -132 -152 -790   24 -254
#>  [659] -306  172  330    8  218   60  234  618  -68 -678  506 -276   18  266
#>  [673] -142  416 -120 -598 -336 -242   66 -292 -578  142 -498 -336 -236 -238
#>  [687]  216 -154 -270  428 -200 -112   82  -82 -384   92  210  194 -216  420
#>  [701]  382  482  -82  -20  100   28  -36 -300   58 -294 -740  282   94  -70
#>  [715]   22  242  -72  530  476  428  222  -54  592   10  242  354 -764  400
#>  [729] -218 -276  198 -142 -702 -204  154  162  364  -28 -334  414 -132 -194
#>  [743] -144   50  534  568  102 -190   10 -140   36 -548 -660  -24  702  148
#>  [757]  328  308  264 -120  -30   36 -324  388 -464  -56  340  236  148 -486
#>  [771]  562  176 -268 -436  622   72 -234 -366 -574   38  196 -558 -304  -66
#>  [785] -312 -948 -232  508 -118 -440 -496 -324 -244 -300  264  348 -180  -78
#>  [799]  -54  214 -242  504 -266  840   56   14  134  126 -170   52  192 -352
#>  [813]  -88 -282 -162  492  166 -258   64  -56 -282 -178  108   66  392  666
#>  [827]  206  204 -160  224 -304  350 -178 -230 -164  286 -366 -698  526 -836
#>  [841]  186  226 -116  306   42 -478  -70  224   44  -10 -428  346   34 -140
#>  [855] -474 -208 -824 -318  -20 -408 -314 -326  -42   74  560 -554   40  782
#>  [869]  120  160 -500  112  358 -100  -56   26 -108  168 -296  416  336 -438
#>  [883] -132 -488  -94  372  128   64 -332  -34   94 -104 -596 -486  442  160
#>  [897]  122  258   72 -660 -294 -366 -206 -492 -324 -648 -228 -468  700  706
#>  [911]  -20  -30  118  274 -140  296  296  680   44  162  -12 -660  434   86
#>  [925] -156 -106  214  320  294  392  410  192   -2   38 -160    0  -62   54
#>  [939]  186  224   50  -78 -346  400  210  322  472 -194 -110 -154  112  212
#>  [953]  254   36  470  482 -200  628 -478  458 -248  -50 -234  124  406  304
#>  [967]  -92   74 -178 -106  402  -30  318  178  552  -70  146 -388 -150 -106
#>  [981]   82  158  484 -404 -584  406  -48 -268   12 -194 -364  626 -430 -416
#>  [995]  478  340 -180 -428  172  -98
#> 
#> $s_statistic
#> [1] 472
#> 
#> $p_value
#> [1] 0.172
#> 
#> $bounds
#>    2.5%   97.5% 
#> -646.05  626.00 
#> 
#> $reject
#> [1] FALSE
#> 
#> $msg
#> [1] "\n - The BBMK test had a p-value of 0.172.\n - At a significance level of 0.05, we fail to reject the null hypothesis.\n - Therefore, there is NO evidence of a trend given the serial correlation."
#>