Sen's Trend Estimator

Computes Sen's linear trend estimator for a univariate time series.

Usage

eda_sens_trend(data, years, quiet = TRUE)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

quiet

Logical scalar. If FALSE, prints a summary of of the statistical test to the console. Default is TRUE.

Value

A list containing the estimated trend:

data: The data argument.

years: The years argument.

slope: Median slope of all pairwise data-year combinations.

intercept: Median intercept estimate of the fitted line.

residuals: Vector of residuals between observed and fitted values.

msg: Character string summarizing the results.

Details

Sen's slope estimator is a robust, nonparametric trend estimator based on the median of all pairwise slopes between data points. The corresponding intercept is the median of each \(y_i - mx_i\) where \(m\) is the estimated slope.

References

Sen, P.K. (1968). Estimates of the regression coefficient based on Kendall's tau. Journal of the American Statistical Association, 63(324), 1379–1389.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
eda_sens_trend(data, years)
#> $data
#>   [1] 106.92862 100.25897 102.30677 113.45502  84.21815  98.97677  82.43465
#>   [8] 103.70060 105.35396  91.20770  97.68624  93.26708 104.92886  79.27116
#>  [15]  97.78326 101.73428 106.31450 103.22489  73.27027  95.80033 103.70320
#>  [22] 102.10024  99.97592  83.25465  99.17297  91.39310 105.19352  93.22492
#>  [29]  99.17705  84.72953 118.55731 103.49558  82.49623  96.23500  98.72704
#>  [36] 105.71291  97.50484 108.02381 117.10142  95.18903  86.38487  80.38076
#>  [43] 103.62647  90.84430 112.93500 109.10819  91.70940  97.33396 118.54191
#>  [50] 125.12024  96.60533 101.08595 114.59925  91.33009  91.55568 100.06430
#>  [57] 112.65379  87.15250  79.90941 100.66470  96.72762  97.51682  85.97496
#>  [64] 101.36356  93.11784 102.07516 104.08187 106.84038 108.13680  95.52265
#>  [71] 121.98109 121.29389 110.50041  71.14720 114.04081  85.53392  97.67344
#>  [78]  90.89083 103.29376  99.21390  98.90397  89.27138 103.95860 108.25451
#>  [85] 108.57277  98.29088 101.00560  95.65229  95.56905  89.66521 122.50497
#>  [92]  94.90634 107.83315 106.31107 111.51340 102.09891  98.66040 107.10639
#>  [99] 102.32650  93.56602
#> 
#> $years
#>   [1] 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915
#>  [16] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#>  [31] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#>  [46] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#>  [61] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#>  [76] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>  [91] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000
#> 
#> $slope
#> [1] 4.073809
#> 
#> $intercept
#> [1] 98.07508
#> 
#> $residuals
#>   [1]   8.81280410   2.10240745   4.10947612  15.21698100 -14.06061752
#>   [6]   0.65726046 -15.92560351   5.29961713   6.91223949  -7.27476579
#>  [11]  -0.83696179  -5.29685538   6.32418254 -19.37425178  -0.90289104
#>  [16]   3.00739133   7.54687039   4.41652710 -25.57883673  -3.08950872
#>  [21]   4.77262151   3.12892459   0.96386544 -15.79814328   0.07943578
#>  [26]  -7.74117533   6.01850770  -5.99082596  -0.07943578 -14.56769291
#>  [31]  19.21934370   4.11687477 -16.92320726  -3.22517479  -0.77387587
#>  [36]   6.17125220  -2.07755547   8.40067587  17.43755213  -4.51557266
#>  [41] -13.36047035 -19.40531874   3.79965319  -9.02325725  13.02670623
#>  [46]   9.15915834  -8.28037423  -2.69654990  18.47065789  25.00825173
#>  [51]  -3.54739108   0.89248606  14.36504662  -8.94484585  -8.76000051
#>  [56]  -0.29211471  12.25663939 -13.28538916 -20.56921799   0.14533092
#>  [61]  -3.83248368  -3.08402654 -14.66662587   0.68123669  -7.60521663
#>  [66]   1.31136327   3.27733533   5.99510320   7.25079120  -5.40409794
#>  [71]  21.01360311  20.28566444   9.45144431 -29.94250111  12.91036975
#>  [76] -15.63725540  -3.53847814 -10.36182353   2.00037202  -2.12023325
#>  [81]  -2.47089906 -12.14422362   2.50225456   6.75742932   7.03494788
#>  [86]  -3.28768030  -0.61369798  -6.00774504  -6.13171757 -12.07630191
#>  [91]  20.72271847  -6.91664208   5.96942836   4.40660696   9.56820195
#>  [96]   0.11296797  -3.36627632   5.03897227   0.21835134  -8.58287229
#> 
#> $msg
#> [1] "Estimated trend: y = 4.074x + 98.08."
#>