White Test for Heteroskedasticity

Performs the White test for heteroskedasticity by regressing the squared residuals of a linear model on the original regressors and their squared terms. The null hypothesis is homoskedasticity.

Usage

eda_white_test(data, years, alpha = 0.05, quiet = TRUE)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

years

Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

quiet

Logical scalar. If FALSE, prints a summary of of the statistical test to the console. Default is TRUE.

Value

A list containing the results of the White test:

data: The data argument.

years: The years argument.

r_squared: Coefficient of determination from the auxiliary regression.

statistic: White test statistic based on sample size and r_squared.

p_value: The p-value derived from a Chi-squared distribution with df = 2.

reject: Logical. If TRUE, the null hypothesis is rejected at alpha.

msg: Character string summarizing the test result, printed if quiet = FALSE.

Details

The White test regresses the squared residuals from a primary linear model lm(data ~ years) against both the original regressor and its square. The test statistic is calculated as \(nR^2\), where \(R^2\) is the coefficient of determination from the auxiliary regression. Under the null hypothesis, the test statistic has the \(\chi^2\) distribution.

References

White, H. (1980). A heteroskedasticity-consistent covariance matrix estimator and a direct test for heteroskedasticity. Econometrica, 48(4), 817–838.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
eda_white_test(data, years)
#> $data
#>   [1] 106.48141 101.00782  82.42012 109.38275  84.40313  97.26097 107.97217
#>   [8]  96.08833 109.67785  93.54223 100.93528 107.62956  94.67642  97.07130
#>  [15] 102.06642  79.20201 108.58147 107.13222  84.83731 109.16774  89.74668
#>  [22]  91.85919 108.46643  94.93178 107.50644 106.77331 101.41293  85.65508
#>  [29]  86.16541  92.84438 122.92140  90.07804 104.59519  97.62281 108.35991
#>  [36]  93.65053  80.76932  98.64659 108.60061 113.31049  83.67603  97.24631
#>  [43]  97.62521 107.06400  99.04726 127.25076 106.30267  88.88680  96.44126
#>  [50] 110.38200  87.75264  93.26501 118.97439  96.53224  78.59880 124.76428
#>  [57]  95.97234 103.79369  92.51198 105.05991  94.64354 103.41252  92.41732
#>  [64]  95.76410  88.48226 115.30293 101.12761  98.74202  85.32868 109.61289
#>  [71] 109.64497 112.58596  84.62194 107.85550 103.97192 106.98846 104.50714
#>  [78] 106.40800 112.20660 110.78431  96.80989  87.26127  93.49333  92.57200
#>  [85]  87.74771  96.87723 110.71671  98.46020  91.09966  84.26416 105.62013
#>  [92]  94.57179 111.64665 103.88534 108.28514 100.32738  93.51255  98.24003
#>  [99]  91.33492 102.79443
#> 
#> $years
#>   [1] 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915
#>  [16] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#>  [31] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#>  [46] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#>  [61] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#>  [76] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>  [91] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000
#> 
#> $r_squared
#> [1] 0.04015465
#> 
#> $statistic
#> [1] 4.015465
#> 
#> $p_value
#> [1] 0.1342928
#> 
#> $reject
#> [1] FALSE
#> 
#> $msg
#> [1] "\n - The White test had a p-value of 0.134.\n - At a significance level of 0.05, we fail to reject the null hypothesis.\n - Therefore, there is NO evidence of heteroskedasticity."
#>