Model Assessment

Computes various metrics for assessing the quality of a fitted flood frequency model. Metrics include accuracy (residual statistics), fitting efficiency (information criteria), and uncertainty (coverage based metrics using confidence intervals).

Usage

model_assessment(
  data,
  distribution,
  params,
  ns_years = NULL,
  ns_structure = NULL,
  alpha = 0.05,
  pp_formula = "Weibull",
  ci = NULL
)

Arguments

data

Numeric vector of observed annual maximum series values. Must be strictly positive, finite, and not missing.

distribution

A three-character code indicating the distribution family. Must be "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

ns_years

For NS-FFA only: Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

ns_structure

For NS-FFA only: Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

alpha

Numeric scalar in \([0.01, 0.1]\). The significance level for confidence intervals or hypothesis tests. Default is 0.05.

pp_formula

Character string specifying the plotting position formula. Must "Weibull" (default), "Blom", "Cunnane", "Gringorten", or "Hazen".

ci

Dataframe containing return periods (in the column periods) and confidence intervals (in the columns ci_lower and ci_upper). Dataframes in this format can be generated with uncertainty_bootstrap(), uncertainty_rfpl(), or uncertainty_rfgpl().

Value

List containing the results of model assessment:

data: The data argument.

estimates: Return level estimates based on plotting positions (S-FFA only).

metrics: A list of model assessment metrics (see details).

Details

These metrics are are computed for both stationary and nonstationary models:

AIC_MLL: Akaike Information Criterion, computed using the maximum log-likelihood.

BIC_MLL: Bayesian Information Criterion, computed using the maximum log-likelihood.

These metrics are always computed for stationary models:

R2: Coefficient of determination from linear regression of estimates vs. data.

RMSE: Root mean squared error of quantile estimates.

bias: Mean bias of quantile estimates.

AIC_RMSE: Akaike Information Criterion, computed using the RMSE.

BIC_RMSE: Bayesian Information Criterion, computed using the RMSE.

These metrics are computed for stationary models if the ci argument is provided:

AW: Average width of the confidence interval(s).

POC: Percent of observations covered by the confidence interval(s).

CWI: Confidence width index, a metric that combines AW and POC.

See also

Examples

# Initialize example data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(100, 10)

# Perform uncertainty analysis
ci <- uncertainty_bootstrap(data, "NOR", "L-moments")$results

# Evaluate model diagnostics
model_assessment(data, "NOR", params, ci = ci)
#> $data
#>   [1] 103.32694  83.67615 110.88718 105.92150  88.93916  97.37292 105.86291
#>   [8]  91.23258 108.51698  99.39284  95.83097 107.19835  85.08089 100.34346
#>  [15]  97.53211 119.54145 100.09658 113.11216  88.64854  95.31387 103.40441
#>  [22]  98.21043 116.15922  99.69659 110.39512 102.20352 109.53315 111.88895
#>  [29]  95.33576 116.22845  92.46341 108.04747  93.51511  94.71189  76.80314
#>  [36] 111.49970  97.32005 113.35798 103.04911 110.48993  92.94706 100.24797
#>  [43] 100.81656 105.06131  94.48316 107.03604 101.50665  81.50408  96.53061
#>  [50]  99.56023  98.79920  92.41629 110.41362 100.44308 108.27265  98.96216
#>  [57]  92.11579 108.29480 116.68956 102.13646 111.57535  89.21120 113.15584
#>  [64]  93.25724  93.56120 100.40112 108.21379 114.65366 104.47294 103.40752
#>  [71] 120.89891 103.25784 105.00181 108.11776  98.44905  91.99071  75.23320
#>  [78]  86.10714 108.21163  99.17586  93.73108  89.23005  93.14664 103.55095
#>  [85] 109.96582 109.94133  89.48270 100.77882 105.86229  96.05944 105.74290
#>  [92] 101.84364  88.81079  88.19487 104.83221 107.28678  90.25614  85.70304
#>  [99]  87.53543  75.85501
#> 
#> $estimates
#>   [1] 123.30079 120.57856 118.85177 117.55301 116.49673 115.59780 114.80973
#>   [8] 114.10420 113.46263 112.87214 112.32341 111.80947 111.32497 110.86568
#>  [15] 110.42824 110.00990 109.60838 109.22178 108.84850 108.48716 108.13657
#>  [22] 107.79571 107.46367 107.13967 106.82300 106.51302 106.20918 105.91097
#>  [29] 105.61792 105.32963 105.04569 104.76577 104.48953 104.21668 103.94693
#>  [36] 103.68003 103.41572 103.15377 102.89397 102.63612 102.38000 102.12543
#>  [43] 101.87224 101.62024 101.36926 101.11915 100.86973 100.62085 100.37236
#>  [50] 100.12409  99.87591  99.62764  99.37915  99.13027  98.88085  98.63074
#>  [57]  98.37976  98.12776  97.87457  97.62000  97.36388  97.10603  96.84623
#>  [64]  96.58428  96.31997  96.05307  95.78332  95.51047  95.23423  94.95431
#>  [71]  94.67037  94.38208  94.08903  93.79082  93.48698  93.17700  92.86033
#>  [78]  92.53633  92.20429  91.86343  91.51284  91.15150  90.77822  90.39162
#>  [85]  89.99010  89.57176  89.13432  88.67503  88.19053  87.67659  87.12786
#>  [92]  86.53737  85.89580  85.19027  84.40220  83.50327  82.44699  81.14823
#>  [99]  79.42144  76.69921
#> 
#> $metrics
#> $metrics$AIC_MLL
#> [1] 740.7242
#> 
#> $metrics$BIC_MLL
#> [1] 745.9345
#> 
#> $metrics$R2
#> [1] 0.9898225
#> 
#> $metrics$RMSE
#> [1] 1.007288
#> 
#> $metrics$bias
#> [1] -0.2654189
#> 
#> $metrics$AIC_RMSE
#> [1] 4.726128
#> 
#> $metrics$BIC_RMSE
#> [1] 9.936468
#> 
#>