L-Distance Method for Distribution Selection

Selects a distribution from a set of candidate distributions by minimizing the Euclidean distance between the theoretical L-moment ratios \((\tau_3, \tau_4)\) and the sample L-moment ratios \((t_3, t_4)\).

Usage

select_ldistance(data)

Arguments

data

Numeric vector of annual maximum series values. Must be strictly positive, finite, and not missing.

Value

A list with the results of distribution selection:

method: "L-distance".

data: The data argument.

metrics: A list of L-distance metrics for each candidate distribution.

recommendation: The name of the distribution with the smallest L-distance.

Details

For each candidate distribution, the method computes the Euclidean distance between sample L-moment ratios (\(\tau_3\), \(\tau_4\)) and the closest point on the theoretical distribution's L-moment curve. For two-parameter distributions (Gumbel, Normal, Log-Normal), the theoretical L-moment ratios are compared directly with the sample L-moment ratios. The distribution with the minimum distance is selected. If a distribution is fit to log-transformed data (Log-Normal or Log-Pearson Type III), the L-moment ratios for the log-transformed sample are used instead.

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
select_ldistance(data)
#> $method
#> [1] "L-distance"
#> 
#> $data
#>   [1] 105.50888  95.11776 108.97312  81.70875 103.55607 111.35757 108.24657
#>   [8]  92.00188  85.72684  95.36425  85.03508  85.35188  89.46015  94.55967
#>  [15] 105.16984 104.29304 106.06805  91.04149 108.23841 104.48857 104.34418
#>  [22]  91.08970  93.56836 114.11590  88.27081  89.45632 104.03920 104.79253
#>  [29] 116.49868  98.45517 141.44174 107.19363 105.30876  95.07604 102.71694
#>  [36]  93.16386 113.56993 115.54153 101.96406  96.39743 113.18202  99.75882
#>  [43] 105.89203 102.38840  98.76708 100.50590  96.98711  95.43674  96.38334
#>  [50]  95.18110  94.10146  90.65543 114.77908 118.29996  90.75872  97.29535
#>  [57]  99.32922  90.32297  82.05011 103.88287  81.25267  86.52339  95.95108
#>  [64] 100.33890  95.51435  96.20210 100.56687 107.18240  98.69194  83.38249
#>  [71] 126.97296 104.12097  93.21544  98.72096 109.38441 117.33438 103.91108
#>  [78] 102.22118 108.38415 101.01798 116.86989  94.62226  91.73635 104.95056
#>  [85] 103.69007  99.91300  93.53077  86.69027 114.98008  92.02169 105.61363
#>  [92] 108.22062 107.05678 101.79175 107.80714  74.17870  85.06983 110.84752
#>  [99] 123.70490  86.89288
#> 
#> $metrics
#> $metrics$GUM
#> [1] 0.1205486
#> 
#> $metrics$NOR
#> [1] 0.05605174
#> 
#> $metrics$LNO
#> [1] 0.02155606
#> 
#> $metrics$GEV
#> [1] 0.03361533
#> 
#> $metrics$GLO
#> [1] 0.0195169
#> 
#> $metrics$GNO
#> [1] 0.02452334
#> 
#> $metrics$PE3
#> [1] 0.0257812
#> 
#> $metrics$LP3
#> [1] 0.02057492
#> 
#> $metrics$WEI
#> [1] 0.04542831
#> 
#> 
#> $recommendation
#> [1] "GLO"
#>