L-Kurtosis Method for Distribution Selection

Selects a probability distribution by minimizing the absolute distance between the theoretical L-kurtosis (\(\tau_4\)) and the sample L-kurtosis (\(t_4\)). Only supports 3-parameter distributions.

For NS-FFA: To select a distribution for a nonstationary model, include the observation years (ns_years) and the nonstationary model structure (ns_structure). Then, this method will detrend the original, nonstationary data internally using the data_decomposition() function prior to distribution selection.

Usage

select_lkurtosis(data, ns_years = NULL, ns_structure = NULL)

Arguments

data

Numeric vector of observed annual maximum series values. Must be strictly positive, finite, and not missing.

ns_years

For NS-FFA only: Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

ns_structure

For NS-FFA only: Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A list with the results of distribution selection:

method: "L-kurtosis".

decomposed_data: The detrended dataset used to compute the L-moments. For S-FFA, this is the data argument. For NS-FFA, it is output of data_decomposition().

metrics: A list of L-kurtosis metrics for each distribution.

recommendation: Name of the distribution with the smallest L-kurtosis metric.

Details

This method computes the distance between the sample and theoretical L-kurtosis values at a fixed L-skewness. For three parameter distributions, the shape parameter that best replicates the sample L-skewness is determined using stats::optim().

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
select_lkurtosis(data)
#> $method
#> [1] "L-kurtosis"
#> 
#> $decomposed_data
#>   [1]  95.18049  96.29055  91.52380 116.75304  93.63964  98.91239  99.26363
#>   [8]  99.44687 103.81568  95.84018  97.27735  97.03057 100.44135  95.40703
#>  [15] 106.52077  89.09580 101.38783 110.83395 100.02482  96.39266  94.80582
#>  [22] 102.89364 113.80124 109.10813  94.03932  96.41940  91.96727  96.95295
#>  [29] 120.58976  87.20401 107.05555  97.25663  98.01527  91.38964  95.79279
#>  [36]  99.98806  87.83996 101.28904 105.78956 102.67137  83.47458  91.94682
#>  [43]  81.61950  99.79259  79.54261 100.15682  93.37544  93.62910  98.02516
#>  [50] 111.46625  85.41590  99.33952 106.29672  88.00825 109.01071 108.62575
#>  [57] 120.74578  93.95437  95.53630  85.72929 109.39748  98.87250 109.97638
#>  [64]  91.37333 108.50989  89.28955  98.41831  78.86985  85.57707  98.92098
#>  [71]  88.35421  93.65519  99.68154  95.85789 118.71775  95.59222 100.94584
#>  [78] 100.27913  89.24481 102.17445 104.32430  98.04964  90.57925 105.59244
#>  [85] 115.97841 106.58272 120.48920  94.44608 102.02191  93.10472 101.73187
#>  [92]  98.61486 104.51093  90.85825 109.25733 103.02943  90.21234  81.61561
#>  [99] 106.38030  92.06038
#> 
#> $metrics
#> $metrics$GEV
#> [1] 0.04347321
#> 
#> $metrics$GLO
#> [1] 0.009128035
#> 
#> $metrics$GNO
#> [1] 0.0351111
#> 
#> $metrics$PE3
#> [1] 0.03690165
#> 
#> $metrics$LP3
#> [1] 0.03651435
#> 
#> $metrics$WEI
#> [1] 0.05714841
#> 
#> 
#> $recommendation
#> [1] "GLO"
#>