Z-Statistic Method for Distribution Selection

Selects the best-fit distribution by comparing a bias-corrected Z-statistic for the sample L-kurtosis (\(\tau_4\)) against the theoretical L-moments for a set of candidate distributions. The distribution with the smallest absolute Z-statistic is selected.

For NS-FFA: To select a distribution for a nonstationary model, include the observation years (ns_years) and the nonstationary model structure (ns_structure). Then, this method will detrend the original, nonstationary data internally using the data_decomposition() function prior to distribution selection.

Usage

select_zstatistic(data, ns_years = NULL, ns_structure = NULL, samples = 10000L)

Arguments

data

Numeric vector of observed annual maximum series values. Must be strictly positive, finite, and not missing.

ns_years

For NS-FFA only: Numeric vector of observation years corresponding to data. Must be the same length as data and strictly increasing.

ns_structure

For NS-FFA only: Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

samples

Integer scalar. The number of bootstrap samples. Default is 10000.

Value

A list with the results of distribution selection:

method: "Z-selection".

decomposed_data: The detrended dataset used to compute the L-moments. For S-FFA, this is the data argument. For NS-FFA, it is output of data_decomposition().

metrics: List of computed Z-statistics for each candidate distribution.

recommendation: Name of the distribution with the smallest Z-statistic.

reg_params: Kappa distribution parameters for the data.

reg_bias_t4: Bias of the L-kurtosis from the bootstrap.

reg_std_t4: Standard deviation of the L-kurtosis from the bootstrap.

log_params: Kappa distribution parameters for the log-transformed data.

log_bias_t4: Bias of the L-kurtosis from the bootstrap using log_params.

log_std_t4: Standard deviation of the L-kurtosis from the bootstrap using log_params.

Details

First, this method fits a four-parameter Kappa distribution to both the original and log-transformed data. Then, bootstrapping is used to estimate the bias and variance of the L-kurtosis. These values, along with the difference between the sample and theoretical L-kurtosis, are used to compute the Z-statistic for each distribution.

References

Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based on L-Moments. Cambridge University Press, New York, USA.

See also

Examples

data <- rnorm(n = 100, mean = 100, sd = 10)
select_zstatistic(data)
#> $method
#> [1] "Z-statistic"
#> 
#> $decomposed_data
#>   [1]  93.66981  90.95956  99.19026  98.71915 101.79338 101.32694 102.04820
#>   [8] 110.98209  99.65298  89.94651  96.69792 108.31218  85.88962 100.37866
#>  [15]  92.39855 102.35374  87.37552  98.86392 107.48181 105.12639 104.11626
#>  [22] 107.28600  91.54723 118.06782 106.62263  99.94238 105.10995  89.14286
#>  [29] 108.08411 104.66837  87.61050 104.62507 112.93330  92.95327  89.89847
#>  [36] 123.78929 102.61699  85.61433 105.39715 116.24928  91.03303  95.75079
#>  [43] 108.26474  94.86348  87.87594  94.51921  83.97493  99.74286  85.59272
#>  [50]  94.79051 112.23648  94.56373 105.61505  99.81454 101.83150  97.17795
#>  [57] 103.62006  91.39505  94.13403 113.16686 103.50459 115.87438 108.12959
#>  [64]  97.86323 107.14469  82.71510  98.38532  94.02230 103.65213  88.34442
#>  [71] 101.53695  83.04860  89.29452 104.70286 100.49462  96.85461  91.09270
#>  [78]  95.05623 108.05659 100.52742 107.24436  82.70252 101.30351 103.20557
#>  [85] 106.39076 109.81046 105.96192  92.45432 102.68295  83.69302 102.26561
#>  [92] 101.81147 104.23127  93.39594  87.95513  83.76912 104.15148 102.00870
#>  [99] 110.92091  93.29045
#> 
#> $metrics
#> $metrics$GEV
#> [1] 0.3596903
#> 
#> $metrics$GLO
#> [1] 2.511239
#> 
#> $metrics$GNO
#> [1] 0.9741756
#> 
#> $metrics$PE3
#> [1] 0.9674902
#> 
#> $metrics$LP3
#> [1] 1.065967
#> 
#> $metrics$WEI
#> [1] 0.5156712
#> 
#> 
#> $recommendation
#> [1] "GEV"
#> 
#> $reg_params
#> [1] 95.82679505  9.75557803  0.36785260  0.09312144
#> 
#> $reg_bias_t4
#> [1] 0.0004975979
#> 
#> $reg_std_t4
#> [1] 0.02868038
#> 
#> $log_params
#> [1] 4.56228252 0.10236845 0.44564137 0.08956027
#> 
#> $log_bias_t4
#> [1] 0.001001428
#> 
#> $log_std_t4
#> [1] 0.02914515
#>