Quantile Functions for Probability Models

Compute the quantiles for stationary and nonstationary probability models.

For NS-FFA: To compute the quantiles for a nonstationary probability model, specify a time slice (ns_slice) and the nonstationary model structure (ns_structure).

Usage

utils_quantiles(p, distribution, params, ns_slice = 0, ns_structure = NULL)

Arguments

p

Numeric vector of probabilities between 0 and 1 with no missing values.

distribution

A three-character code indicating the distribution family. Must be "GUM", "NOR", "LNO", "GEV", "GLO", "GNO", "PE3", "LP3", or "WEI".

params

Numeric vector of distribution parameters, in the order (location, scale, shape). The length must be between 2 and 5, depending on the specified distribution and structure.

ns_slice

For NS-FFA only: Numeric scalar specifying the year at which to evaluate the quantiles of a nonstationary probability distribution. ns_slice does not have to be an element of the ns_years argument.

ns_structure

For NS-FFA only: Named list indicating which distribution parameters are modeled as nonstationary. Must contain two logical scalars:

location: If TRUE, the location parameter has a linear temporal trend.

scale: If TRUE, the scale parameter has a linear temporal trend.

Value

A numeric vector of quantiles with the same length as p.

Examples

p <- runif(n = 100)
params <- c(1, 1, 1)
utils_quantiles(p, "GEV", params)
#>   [1]  2.4666847  3.4950121 26.1556802  3.2345470  2.6931867  0.2435830
#>   [7]  0.6269835  2.8402128 10.2380081  3.6969446  1.1312619  1.9922922
#>  [13]  1.4325774  0.8195452  1.0835757  1.2036872  2.8097046  0.6565409
#>  [19]  0.4012199  0.3121710  1.5241493 21.6154858  0.7465035  1.9976075
#>  [25]  1.2301924  6.3474841  1.3096856  5.6510131  4.8775172  1.5102954
#>  [31]  3.3410673  2.0313215  0.1723445  3.2281077  2.9360130  0.7143273
#>  [37]  1.3371532  0.8591236  0.6466655  0.3833000  0.6599588  0.9787352
#>  [43]  0.2989051  0.5866714  1.5104043  1.6610655 11.0243983  1.3211304
#>  [49]  0.2959296  1.3172886  7.4973113  1.9440010  0.8659400  0.6121291
#>  [55] 47.2800933  2.2525122  0.9202294  2.1745096  4.1532640  1.4217814
#>  [61]  0.6584235  0.8220539  0.6115734  0.4342235  1.3889402  1.2784336
#>  [67] 18.9807630  0.4587727  1.2858622  1.1816469 39.8984444  0.8202021
#>  [73]  0.9376804  1.1062952  5.9285384  0.9373009  1.9164821  0.6941272
#>  [79]  1.1921584  0.6312081  1.3851441  0.5346806  2.6501865  0.9666124
#>  [85]  1.3264494  6.5310088 12.6796871  1.9229663 28.1915678  1.4339464
#>  [91]  0.6017530  0.6988725  0.7521211  1.6447825  1.1599285  0.9013736
#>  [97]  0.7129161 12.1867549  0.6787519  0.9381461